+++ db2/history2/reviewable_request_details.go
@@ -13,21 +13,19 @@ import (
 
 //ReviewableRequestDetails - stores in union switch details of the reviewable requests
 type ReviewableRequestDetails struct {
-	Type                     xdr.ReviewableRequestType `json:"type"`
-	AssetCreation            *AssetCreationRequest     `json:"asset_create,omitempty"`
-	AssetUpdate              *AssetUpdateRequest       `json:"asset_update,omitempty"`
-	PreIssuanceCreate        *PreIssuanceRequest       `json:"pre_issuance_create,omitempty"`
-	IssuanceCreate           *IssuanceRequest          `json:"issuance_create,omitempty"`
-	Withdraw                 *WithdrawalRequest        `json:"withdraw,omitempty"`
-	TwoStepWithdraw          *WithdrawalRequest        `json:"two_step_withdrawal"`
-	Sale                     *SaleRequest              `json:"sale,omitempty"`
-	LimitsUpdate             *LimitsUpdateRequest      `json:"limits_update"`
-	AmlAlert                 *AmlAlertRequest          `json:"aml_alert"`
-	ChangeRole               *ChangeRoleRequest        `json:"change_role,omitempty"`
-	UpdateSaleDetails        *UpdateSaleDetailsRequest `json:"update_sale_details"`
-	UpdateSaleEndTimeRequest *UpdateSaleEndTimeRequest `json:"update_sale_end_time_request"`
-	AtomicSwapBidCreation    *AtomicSwapBidCreation    `json:"atomic_swap_bid_creation"`
-	AtomicSwap               *AtomicSwap               `json:"atomic_swap"`
+	Type                xdr.ReviewableRequestType   `json:"type"`
+	CreateAsset         *CreateAssetRequest         `json:"create_asset,omitempty"`
+	UpdateAsset         *UpdateAssetRequest         `json:"update_asset,omitempty"`
+	CreatePreIssuance   *CreatePreIssuanceRequest   `json:"create_pre_issuance,omitempty"`
+	CreateIssuance      *CreateIssuanceRequest      `json:"create_issuance,omitempty"`
+	CreateWithdraw      *CreateWithdrawalRequest    `json:"create_withdraw,omitempty"`
+	CreateSale          *CreateSaleRequest          `json:"create_sale,omitempty"`
+	UpdateLimits        *UpdateLimitsRequest        `json:"update_limits"`
+	CreateAmlAlert      *CreateAmlAlertRequest      `json:"create_aml_alert"`
+	ChangeRole          *ChangeRoleRequest          `json:"change_role,omitempty"`
+	UpdateSaleDetails   *UpdateSaleDetailsRequest   `json:"update_sale_details"`
+	CreateAtomicSwapBid *CreateAtomicSwapBidRequest `json:"create_atomic_swap_bid"`
+	CreateAtomicSwap    *CreateAtomicSwapRequest    `json:"create_atomic_swap"`
 }
 
 //Value - implements db driver method for auto marshal
@@ -50,74 +48,75 @@ func (r *ReviewableRequestDetails) Scan(src interface{}) error {
 	return nil
 }
 
-//AssetCreationRequest - asset creation request details
-type AssetCreationRequest struct {
+//CreateAssetRequest - asset creation request details
+type CreateAssetRequest struct {
 	Asset                  string            `json:"asset"`
 	Policies               int32             `json:"policies"`
 	PreIssuedAssetSigner   string            `json:"pre_issued_asset_signer"`
 	MaxIssuanceAmount      string            `json:"max_issuance_amount"`
 	InitialPreissuedAmount string            `json:"initial_preissued_amount"`
-	Details                regources.Details `json:"details"`
+	CreatorDetails         regources.Details `json:"details"`
 }
 
-//AssetUpdateRequest - asset update request details
-type AssetUpdateRequest struct {
-	Asset    string            `json:"asset"`
-	Policies int32             `json:"policies"`
-	Details  regources.Details `json:"details"`
+//UpdateAssetRequest - asset update request details
+type UpdateAssetRequest struct {
+	Asset          string            `json:"asset"`
+	Policies       int32             `json:"policies"`
+	CreatorDetails regources.Details `json:"details"`
 }
 
-//PreIssuanceRequest - request details
-type PreIssuanceRequest struct {
-	Asset     string `json:"asset"`
-	Amount    string `json:"amount"`
-	Signature string `json:"signature"`
-	Reference string `json:"reference"`
+//CreatePreIssuanceRequest - request details
+type CreatePreIssuanceRequest struct {
+	Asset          string            `json:"asset"`
+	Amount         string            `json:"amount"`
+	Signature      string            `json:"signature"`
+	Reference      string            `json:"reference"`
+	CreatorDetails regources.Details `json:"creator_details"`
 }
 
-//IssuanceRequest - request details
-type IssuanceRequest struct {
-	Asset    string            `json:"asset"`
-	Amount   string            `json:"amount"`
-	Receiver string            `json:"receiver"`
-	Details  regources.Details `json:"external_details"`
+//CreateIssuanceRequest - request details
+type CreateIssuanceRequest struct {
+	Asset          string            `json:"asset"`
+	Amount         string            `json:"amount"`
+	Receiver       string            `json:"receiver"`
+	CreatorDetails regources.Details `json:"external_details"`
 }
 
-//WithdrawalRequest - request details
-type WithdrawalRequest struct {
+//CreateWithdrawalRequest - request details
+type CreateWithdrawalRequest struct {
 	BalanceID       string            `json:"balance_id"`
 	Amount          string            `json:"amount"`
 	FixedFee        string            `json:"fixed_fee"`
 	PercentFee      string            `json:"percent_fee"`
-	Details         regources.Details `json:"external_details"`
+	CreatorDetails  regources.Details `json:"creator_details"`
 	ReviewerDetails regources.Details `json:"reviewer_details"`
 }
 
-//SaleRequest - request details
-type SaleRequest struct {
+//CreateSaleRequest - request details
+type CreateSaleRequest struct {
 	BaseAsset           string                 `json:"base_asset"`
 	DefaultQuoteAsset   string                 `json:"quote_asset"`
 	StartTime           time.Time              `json:"start_time"`
 	EndTime             time.Time              `json:"end_time"`
 	SoftCap             string                 `json:"soft_cap"`
 	HardCap             string                 `json:"hard_cap"`
-	Details             regources.Details      `json:"details"`
+	CreatorDetails      regources.Details      `json:"creator_details"`
 	QuoteAssets         []regources.AssetPrice `json:"quote_assets"`
 	SaleType            xdr.SaleType           `json:"sale_type"`
 	BaseAssetForHardCap string                 `json:"base_asset_for_hard_cap"`
 }
 
-//LimitsUpdateRequest - request details
-type LimitsUpdateRequest struct {
-	DocumentHash string            `json:"document_hash"`
-	Details      regources.Details `json:"details"`
+//UpdateLimitsRequest - request details
+type UpdateLimitsRequest struct {
+	DocumentHash   string            `json:"document_hash"`
+	CreatorDetails regources.Details `json:"creator_details"`
 }
 
-//AmlAlertRequest - request details
-type AmlAlertRequest struct {
-	BalanceID string `json:"balance_id"`
-	Amount    string `json:"amount"`
-	Reason    string `json:"reason"`
+//CreateAmlAlertRequest - request details
+type CreateAmlAlertRequest struct {
+	BalanceID      string `json:"balance_id"`
+	Amount         string `json:"amount"`
+	CreatorDetails string `json:"creator_details"`
 }
 
 //ChangeRoleRequest - request details
@@ -126,31 +125,27 @@ type ChangeRoleRequest struct {
 	AccountRoleToSet   uint64            `json:"account_role_to_set"`
 	KYCData            regources.Details `json:"kyc_data"`
 	SequenceNumber     uint32            `json:"sequence_number"`
+	CreatorDetails     regources.Details `json:"creator_details"`
 }
 
 //UpdateSaleDetailsRequest - request details
 type UpdateSaleDetailsRequest struct {
-	SaleID     uint64            `json:"sale_id"`
-	NewDetails regources.Details `json:"new_details"`
+	SaleID         uint64            `json:"sale_id"`
+	CreatorDetails regources.Details `json:"creator_details"`
 }
 
-//UpdateSaleEndTimeRequest - request details
-type UpdateSaleEndTimeRequest struct {
-	SaleID     uint64    `json:"sale_id"`
-	NewEndTime time.Time `json:"new_end_time"`
+//CreateAtomicSwapBidRequest - request details
+type CreateAtomicSwapBidRequest struct {
+	BaseBalance    string                 `json:"base_balance"`
+	BaseAmount     uint64                 `json:"base_amount"`
+	CreatorDetails regources.Details      `json:"creator_details"`
+	QuoteAssets    []regources.AssetPrice `json:"quote_assets"`
 }
 
-//AtomicSwapBidCreation - request details
-type AtomicSwapBidCreation struct {
-	BaseBalance string                 `json:"base_balance"`
-	BaseAmount  uint64                 `json:"base_amount"`
-	Details     regources.Details      `json:"details"`
-	QuoteAssets []regources.AssetPrice `json:"quote_assets"`
-}
-
-//AtomicSwap - request details
-type AtomicSwap struct {
-	BidID      uint64 `json:"bid_id"`
-	BaseAmount uint64 `json:"base_amount"`
-	QuoteAsset string `json:"quote_asset"`
+//CreateAtomicSwapRequest - request details
+type CreateAtomicSwapRequest struct {
+	BidID          uint64            `json:"bid_id"`
+	BaseAmount     uint64            `json:"base_amount"`
+	QuoteAsset     string            `json:"quote_asset"`
+	CreatorDetails regources.Details `json:"creator_details"`
 }