+++ web_v2/resources/request_details.go
@@ -11,30 +11,30 @@ import (
 // NewRequestDetails - returns new instance of reviewable request details
 func NewRequestDetails(request history2.ReviewableRequest) regources.Resource {
 	switch request.RequestType {
-	case xdr.ReviewableRequestTypeAssetUpdate:
-		return newAssetUpdateRequest(request.ID, *request.Details.AssetUpdate)
-	case xdr.ReviewableRequestTypePreIssuanceCreate:
-		return newPreIssuanceRequest(request.ID, *request.Details.PreIssuanceCreate)
-	case xdr.ReviewableRequestTypeIssuanceCreate:
-		return newIssuanceRequest(request.ID, *request.Details.IssuanceCreate)
-	case xdr.ReviewableRequestTypeWithdraw:
-		return newWithdrawalRequest(request.ID, *request.Details.Withdraw)
-	case xdr.ReviewableRequestTypeSale:
-		return newSaleRequest(request.ID, *request.Details.Sale)
-	case xdr.ReviewableRequestTypeLimitsUpdate:
-		return newLimitsUpdateRequest(request.ID, *request.Details.LimitsUpdate)
-	case xdr.ReviewableRequestTypeAmlAlert:
-		return newAmlAlertRequest(request.ID, *request.Details.AmlAlert)
+	case xdr.ReviewableRequestTypeUpdateAsset:
+		return newAssetUpdateRequest(request.ID, *request.Details.UpdateAsset)
+	case xdr.ReviewableRequestTypeCreatePreIssuance:
+		return newPreIssuanceRequest(request.ID, *request.Details.CreatePreIssuance)
+	case xdr.ReviewableRequestTypeCreateIssuance:
+		return newIssuanceRequest(request.ID, *request.Details.CreateIssuance)
+	case xdr.ReviewableRequestTypeCreateWithdraw:
+		return newWithdrawalRequest(request.ID, *request.Details.CreateWithdraw)
+	case xdr.ReviewableRequestTypeCreateSale:
+		return newSaleRequest(request.ID, *request.Details.CreateSale)
+	case xdr.ReviewableRequestTypeUpdateLimits:
+		return newLimitsUpdateRequest(request.ID, *request.Details.UpdateLimits)
+	case xdr.ReviewableRequestTypeCreateAmlAlert:
+		return newAmlAlertRequest(request.ID, *request.Details.CreateAmlAlert)
 	case xdr.ReviewableRequestTypeChangeRole:
 		return newChangeRoleRequest(request.ID, *request.Details.ChangeRole)
 	case xdr.ReviewableRequestTypeUpdateSaleDetails:
 		return newUpdateSaleDetailsRequest(request.ID, *request.Details.UpdateSaleDetails)
-	case xdr.ReviewableRequestTypeAssetCreate:
-		return newAssetCreateRequest(request.ID, *request.Details.AssetCreation)
+	case xdr.ReviewableRequestTypeCreateAsset:
+		return newAssetCreateRequest(request.ID, *request.Details.CreateAsset)
 	case xdr.ReviewableRequestTypeCreateAtomicSwapBid:
-		return newAtomicSwapBidRequest(request.ID, *request.Details.AtomicSwapBidCreation)
-	case xdr.ReviewableRequestTypeAtomicSwap:
-		return newAtomicSwapRequest(request.ID, *request.Details.AtomicSwap)
+		return newAtomicSwapBidRequest(request.ID, *request.Details.CreateAtomicSwapBid)
+	case xdr.ReviewableRequestTypeCreateAtomicSwap:
+		return newAtomicSwapRequest(request.ID, *request.Details.CreateAtomicSwap)
 	default:
 		panic(errors.From(errors.New("unexpected operation type"), logan.F{
 			"type": request.RequestType,
@@ -43,56 +43,56 @@ func NewRequestDetails(request history2.ReviewableRequest) regources.Resource {
 	return nil
 }
 
-func newAmlAlertRequest(id int64, details history2.AmlAlertRequest) *regources.AmlAlertRequest {
-	return &regources.AmlAlertRequest{
+func newAmlAlertRequest(id int64, details history2.CreateAmlAlertRequest) *regources.CreateAmlAlertRequest {
+	return &regources.CreateAmlAlertRequest{
 		Key: regources.NewKeyInt64(id, regources.TypeRequestDetailsAMLAlert),
-		Attributes: regources.AmlAlertRequestAttrs{
-			Amount: details.Amount,
-			Reason: details.Reason,
+		Attributes: regources.CreateAmlAlertRequestAttrs{
+			Amount:         details.Amount,
+			CreatorDetails: details.CreatorDetails,
 		},
-		Relationships: regources.AmlAlertRequestRelations{
+		Relationships: regources.CreateAmlAlertRequestRelations{
 			Balance: NewBalanceKey(details.BalanceID).AsRelation(),
 		},
 	}
 }
-func newAssetCreateRequest(id int64, details history2.AssetCreationRequest) *regources.AssetCreateRequest {
-	return &regources.AssetCreateRequest{
+func newAssetCreateRequest(id int64, details history2.CreateAssetRequest) *regources.CreateAssetRequest {
+	return &regources.CreateAssetRequest{
 		Key: regources.NewKeyInt64(id, regources.TypeRequestDetailsAssetCreate),
-		Attributes: regources.AssetCreateRequestAttrs{
+		Attributes: regources.CreateAssetRequestAttrs{
 			Asset:                  details.Asset,
 			Policies:               details.Policies,
 			PreIssuanceAssetSigner: details.PreIssuedAssetSigner,
 			MaxIssuanceAmount:      details.MaxIssuanceAmount,
 			InitialPreissuedAmount: details.InitialPreissuedAmount,
-			Details:                details.Details,
+			CreatorDetails:         details.CreatorDetails,
 		},
 	}
 }
-func newAssetUpdateRequest(id int64, details history2.AssetUpdateRequest) *regources.AssetUpdateRequest {
-	return &regources.AssetUpdateRequest{
+func newAssetUpdateRequest(id int64, details history2.UpdateAssetRequest) *regources.UpdateAssetRequest {
+	return &regources.UpdateAssetRequest{
 		Key: regources.NewKeyInt64(id, regources.TypeRequestDetailsAssetUpdate),
 		Attributes: regources.AssetUpdateRequestAttrs{
-			Policies: details.Policies,
-			Details:  details.Details,
+			Policies:       details.Policies,
+			CreatorDetails: details.CreatorDetails,
 		},
-		Relationships: regources.AssetUpdateRequestRelations{
+		Relationships: regources.UpdateAssetRequestRelations{
 			Asset: NewAssetKey(details.Asset).AsRelation(),
 		},
 	}
 }
-func newAtomicSwapRequest(id int64, details history2.AtomicSwap) *regources.AtomicSwapRequest {
-	return &regources.AtomicSwapRequest{
+func newAtomicSwapRequest(id int64, details history2.CreateAtomicSwapRequest) *regources.CreateAtomicSwapRequest {
+	return &regources.CreateAtomicSwapRequest{
 		Key: regources.NewKeyInt64(id, regources.TypeRequestDetailsAtomicSwap),
-		Attributes: regources.AtomicSwapRequestAttrs{
+		Attributes: regources.CreateAtomicSwapRequestAttrs{
 			BaseAmount: regources.Amount(details.BaseAmount),
 		},
-		Relationships: regources.AtomicSwapRequestRelations{
+		Relationships: regources.CreateAtomicSwapRequestRelations{
 			Bid:        regources.NewKeyInt64(int64(details.BidID), regources.TypeAswapBid).AsRelation(),
 			QuoteAsset: newQuoteAssetKey(details.QuoteAsset).AsRelation(),
 		},
 	}
 }
-func newAtomicSwapBidRequest(id int64, details history2.AtomicSwapBidCreation) *regources.AtomicSwapBidRequest {
+func newAtomicSwapBidRequest(id int64, details history2.CreateAtomicSwapBidRequest) *regources.CreateAtomicSwapBidRequest {
 	quoteAssets := &regources.RelationCollection{
 		Data: make([]regources.Key, 0, len(details.QuoteAssets)),
 	}
@@ -100,54 +100,54 @@ func newAtomicSwapBidRequest(id int64, details history2.AtomicSwapBidCreation) *
 		quoteAssets.Data = append(quoteAssets.Data, newQuoteAssetKey(quoteAsset.Asset))
 	}
 
-	return &regources.AtomicSwapBidRequest{
+	return &regources.CreateAtomicSwapBidRequest{
 		Key: regources.NewKeyInt64(id, regources.TypeRequestDetailsAtomicSwapBid),
-		Attributes: regources.AtomicSwapBidRequestAttrs{
-			BaseAmount: regources.Amount(details.BaseAmount),
-			Details:    details.Details,
+		Attributes: regources.CreateAtomicSwapBidRequestAttrs{
+			BaseAmount:     regources.Amount(details.BaseAmount),
+			CreatorDetails: details.CreatorDetails,
 		},
-		Relationships: regources.AtomicSwapBidRequestRelations{
+		Relationships: regources.CreateAtomicSwapBidRequestRelations{
 			BaseBalance: NewBalanceKey(details.BaseBalance).AsRelation(),
 			QuoteAssets: quoteAssets,
 		},
 	}
 }
-func newIssuanceRequest(id int64, details history2.IssuanceRequest) *regources.IssuanceRequest {
-	return &regources.IssuanceRequest{
+func newIssuanceRequest(id int64, details history2.CreateIssuanceRequest) *regources.CreateIssuanceRequest {
+	return &regources.CreateIssuanceRequest{
 		Key: regources.NewKeyInt64(id, regources.TypeRequestDetailsIssuance),
-		Attributes: regources.IssuanceRequestAttrs{
-			Amount:  details.Amount,
-			Details: details.Details,
+		Attributes: regources.CreateIssuanceRequestAttrs{
+			Amount:         details.Amount,
+			CreatorDetails: details.CreatorDetails,
 		},
-		Relationships: regources.IssuanceRequestRelations{
+		Relationships: regources.CreateIssuanceRequestRelations{
 			Asset:    NewAssetKey(details.Asset).AsRelation(),
 			Receiver: NewBalanceKey(details.Receiver).AsRelation(),
 		},
 	}
 }
-func newLimitsUpdateRequest(id int64, details history2.LimitsUpdateRequest) *regources.LimitsUpdateRequest {
-	return &regources.LimitsUpdateRequest{
+func newLimitsUpdateRequest(id int64, details history2.UpdateLimitsRequest) *regources.UpdateLimitsRequest {
+	return &regources.UpdateLimitsRequest{
 		Key: regources.NewKeyInt64(id, regources.TypeRequestDetailsLimitsUpdate),
-		Attributes: regources.LimitsUpdateRequestAttrs{
-			DocumentHash: details.DocumentHash,
-			Details:      details.Details,
+		Attributes: regources.UpdateLimitsRequestAttrs{
+			DocumentHash:   details.DocumentHash,
+			CreatorDetails: details.CreatorDetails,
 		},
 	}
 }
-func newPreIssuanceRequest(id int64, details history2.PreIssuanceRequest) *regources.PreIssuanceRequest {
-	return &regources.PreIssuanceRequest{
+func newPreIssuanceRequest(id int64, details history2.CreatePreIssuanceRequest) *regources.CreatePreIssuanceRequest {
+	return &regources.CreatePreIssuanceRequest{
 		Key: regources.NewKeyInt64(id, regources.TypeRequestDetailsPreIssuance),
-		Attributes: regources.PreIssuanceRequestAttrs{
+		Attributes: regources.CreatePreIssuanceRequestAttrs{
 			Amount:    details.Amount,
 			Signature: details.Signature,
 			Reference: details.Reference,
 		},
-		Relationships: regources.PreIssuanceRequestRelations{
+		Relationships: regources.CreatePreIssuanceRequestRelations{
 			Asset: NewAssetKey(details.Asset).AsRelation(),
 		},
 	}
 }
-func newSaleRequest(id int64, details history2.SaleRequest) *regources.SaleRequest {
+func newSaleRequest(id int64, details history2.CreateSaleRequest) *regources.CreateSaleRequest {
 	quoteAssets := &regources.RelationCollection{
 		Data: make([]regources.Key, 0, len(details.QuoteAssets)),
 	}
@@ -155,16 +155,16 @@ func newSaleRequest(id int64, details history2.SaleRequest) *regources.SaleReque
 		quoteAssets.Data = append(quoteAssets.Data, newQuoteAssetKey(quoteAsset.Asset))
 	}
 
-	return &regources.SaleRequest{
+	return &regources.CreateSaleRequest{
 		Key: regources.NewKeyInt64(id, regources.TypeRequestDetailsSale),
-		Attributes: regources.SaleRequestAttrs{
+		Attributes: regources.CreateSaleRequestAttrs{
 			BaseAssetForHardCap: details.BaseAssetForHardCap,
 			StartTime:           details.StartTime,
 			EndTime:             details.EndTime,
 			SaleType:            details.SaleType,
-			Details:             details.Details,
+			CreatorDetails:      details.CreatorDetails,
 		},
-		Relationships: regources.SaleRequestRelations{
+		Relationships: regources.CreateSaleRequestRelations{
 			QuoteAssets:       quoteAssets,
 			BaseAsset:         NewAssetKey(details.BaseAsset).AsRelation(),
 			DefaultQuoteAsset: newQuoteAssetKey(details.DefaultQuoteAsset).AsRelation(),
@@ -188,26 +188,26 @@ func newUpdateSaleDetailsRequest(id int64, details history2.UpdateSaleDetailsReq
 	return &regources.UpdateSaleDetailsRequest{
 		Key: regources.NewKeyInt64(id, regources.TypeRequestDetailsUpdateSaleDetails),
 		Attributes: regources.UpdateSaleDetailsRequestAttrs{
-			NewDetails: details.NewDetails,
+			CreatorDetails: details.CreatorDetails,
 		},
 		Relationships: regources.UpdateSaleDetailsRequestRelations{
 			Sale: NewSaleKey(int64(details.SaleID)).AsRelation(),
 		},
 	}
 }
-func newWithdrawalRequest(id int64, details history2.WithdrawalRequest) *regources.WithdrawalRequest {
-	return &regources.WithdrawalRequest{
+func newWithdrawalRequest(id int64, details history2.CreateWithdrawalRequest) *regources.CreateWithdrawalRequest {
+	return &regources.CreateWithdrawalRequest{
 		Key: regources.NewKeyInt64(id, regources.TypeRequestDetailsWithdrawal),
-		Attributes: regources.WithdrawalRequestAttrs{
+		Attributes: regources.CreateWithdrawalRequestAttrs{
 			Fee: regources.FeeStr{
 				Fixed:             details.FixedFee,
 				CalculatedPercent: details.PercentFee,
 			},
 			Amount:          details.Amount,
-			Details:         details.Details,
+			CreatorDetails:  details.CreatorDetails,
 			ReviewerDetails: details.ReviewerDetails,
 		},
-		Relationships: regources.WithdrawalRequestRelations{
+		Relationships: regources.CreateWithdrawalRequestRelations{
 			Balance: NewBalanceKey(details.BalanceID).AsRelation(),
 		},
 	}